# Backend/src/docs/swagger.yaml
definitions:
  model.ItemPictures:
    properties:
      id:
        type: string
      itemID:
        type: string
      picture:
        type: string
    type: object
  model.ItemSubjects:
    properties:
      id:
        type: string
      itemID:
        type: string
      subjectID:
        type: string
    type: object
  model.KeywordsForItems:
    properties:
      id:
        type: string
      itemID:
        type: string
      keyword:
        type: string
    type: object
  model.Rooms:
    properties:
      id:
        type: string
      name:
        type: string
      warehouseID:
        type: string
    type: object
  model.Warehouses:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.INVErrorMessage:
    properties:
      errorMessage:
        type: string
    type: object
  models.ItemWithEverything:
    properties:
      classFour:
        type: string
      classOne:
        type: string
      classThree:
        type: string
      classTwo:
        type: string
      damaged:
        type: boolean
      damagedDesc:
        type: string
      description:
        type: string
      id:
        type: string
      keywords:
        items:
          $ref: '#/definitions/model.KeywordsForItems'
        type: array
      name:
        type: string
      pictures:
        items:
          $ref: '#/definitions/model.ItemPictures'
        type: array
      quantity:
        type: integer
      status:
        type: string
      subject:
        items:
          $ref: '#/definitions/model.ItemSubjects'
        type: array
    type: object
  models.OwnShelve:
    properties:
      id:
        type: string
      roomID:
        type: string
      shelveTypeName:
        type: string
    type: object
  models.RoomWithShelves:
    properties:
      id:
        type: string
      name:
        type: string
      shelves:
        items:
          properties:
            id:
              type: string
            roomID:
              type: string
            shelveTypeID:
              type: string
          type: object
        type: array
      warehouseID:
        type: string
    type: object
  models.ShelveWithItems:
    properties:
      id:
        type: string
      items:
        items:
          properties:
            classFour:
              type: boolean
            classOne:
              type: boolean
            classThree:
              type: boolean
            classTwo:
              type: boolean
            damaged:
              type: boolean
            damagedDescription:
              type: string
            description:
              type: string
            id:
              type: string
            name:
              type: string
            quantity:
              type: integer
            statusID:
              type: string
          type: object
        type: array
      roomID:
        type: string
      shelveTypeName:
        type: string
    type: object
  models.WarehouseWithRooms:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      rooms:
        items:
          properties:
            id:
              type: string
            name:
              type: string
            warehouseID:
              type: string
          type: object
        type: array
    type: object
info:
  contact: {}
paths:
  /items:
    get:
      consumes:
      - application/json
      description: Get items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ItemWithEverything'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Get items
      tags:
      - Items
  /rooms:
    get:
      consumes:
      - application/json
      description: Get rooms
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Rooms'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Get rooms
      tags:
      - Rooms
    post:
      consumes:
      - application/json
      description: Create Room
      parameters:
      - description: Room model
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/model.Rooms'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Rooms'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Create Room
      tags:
      - Rooms
    put:
      consumes:
      - application/json
      description: Update room
      parameters:
      - description: Room model
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/model.Rooms'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Rooms'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Update room
      tags:
      - Rooms
  /rooms/{id}:
    get:
      consumes:
      - application/json
      description: Get room by id with shelves
      parameters:
      - description: Room id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoomWithShelves'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Get room by id with shelves
      tags:
      - Rooms
  /shelves:
    get:
      consumes:
      - application/json
      description: Get shelves
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OwnShelve'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Get shelves
      tags:
      - Shelves
  /shelves/{id}:
    get:
      consumes:
      - application/json
      description: Get shelve by id
      parameters:
      - description: Shelve id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OwnShelve'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Get shelve by id
      tags:
      - Shelves
  /shelveswithitems:
    get:
      consumes:
      - application/json
      description: Get shelves with items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShelveWithItems'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Get shelves with items
      tags:
      - Shelves
  /shelveswithitems/{id}:
    get:
      consumes:
      - application/json
      description: Get shelve by id with items
      parameters:
      - description: shelve id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShelveWithItems'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Get shelve by id with items
      tags:
      - Shelves
  /warehouses:
    get:
      consumes:
      - application/json
      description: Get warehouses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Warehouses'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Get warehouses
      tags:
      - Warehouses
    post:
      consumes:
      - application/json
      description: Create warehouse
      parameters:
      - description: Warehouse name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Warehouses'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Create warehouse
      tags:
      - Warehouses
    put:
      consumes:
      - application/json
      description: Update warehouse
      parameters:
      - description: Warehouses model
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/model.Warehouses'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Warehouses'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Update warehouse
      tags:
      - Warehouses
  /warehouses/{id}:
    get:
      consumes:
      - application/json
      description: Get warehouse by id
      parameters:
      - description: Warehouse id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.WarehouseWithRooms'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.INVErrorMessage'
      summary: Get warehouse by id
      tags:
      - Warehouses
swagger: "2.0"